adding more view tests and moving mixins to their own fiel
